Extract HarvestResource into own file and cache trait lookups in constructor
